some renaming, added large comment to globals
